Revert "Use unsigned for capturing groups in Regex interfaces"
This reverts commit bbf8fb95ddee9ec9ebed30ca266cdc4c12ad094f.  The earlier commit inadvertently
introduced the possibility of a uint16 overflow in the regex parser, so back that out before we
release it and then go back and look at improving both the internal interface and the form
factor for reporting the too-many-capturing-groups error to the user.

+++ lib/Parser/RegexPattern.h
@@ -58,7 +58,7 @@ namespace UnifiedRegex
         Js::ScriptContext *GetScriptContext() const;
 
         inline bool IsLiteral() const { return isLiteral; }
-        uint16 NumGroups() const;
+        int NumGroups() const;
         bool IsIgnoreCase() const;
         bool IsGlobal() const;
         bool IsMultiline() const;
